+++ apps/desktop/src/components/assistant-ui/thread/block-direction.test.tsx
@@ -10,7 +10,7 @@ import { AssistantRuntimeProvider, type ThreadMessage, useExternalStoreRuntime }
 import { render, screen } from '@testing-library/react'
 import { describe, expect, it, vi } from 'vitest'
 
-import { Thread } from './thread'
+import { Thread } from '.'
 
 const createdAt = new Date('2026-06-01T00:00:00.000Z')
