+++ apps/desktop/src/components/assistant-ui/thread/content.ts
