+++ apps/desktop/src/components/assistant-ui/thread/status.tsx
