+++ apps/desktop/src/components/assistant-ui/thread/system-message.tsx
@@ -1,7 +1,7 @@
 import { MessagePrimitive, useAuiState } from '@assistant-ui/react'
 import { type FC } from 'react'
 
-import { messageContentText } from '@/components/assistant-ui/thread-content'
+import { messageContentText } from '@/components/assistant-ui/thread/content'
 import { Codicon } from '@/components/ui/codicon'
 import { LinkifiedText } from '@/lib/external-link'
 import { cn } from '@/lib/utils'
