+++ apps/desktop/src/components/assistant-ui/thread/timeline-data.ts
