+++ apps/desktop/src/components/assistant-ui/thread/timestamp.ts
