+++ apps/desktop/src/components/assistant-ui/thread/types.ts
