+++ apps/desktop/src/components/assistant-ui/thread/user-message-text.tsx
