+++ apps/desktop/src/components/assistant-ui/tool/fallback-model.ts
