+++ apps/desktop/src/components/prompt-overlays.tsx
@@ -3,7 +3,7 @@
 import { useStore } from '@nanostores/react'
 import { type FormEvent, useCallback, useEffect, useState } from 'react'
 
-import { PendingApprovalFallback } from '@/components/assistant-ui/tool-approval'
+import { PendingApprovalFallback } from '@/components/assistant-ui/tool/approval'
 import { Button } from '@/components/ui/button'
 import {
   Dialog,
